Implement l10n from tar
