+++ crates/zng-ext-l10n/Cargo.toml
@@ -11,6 +11,11 @@ repository = "https://github.com/zng-ui/zng"
 categories = ["gui"]
 keywords = ["gui", "ui", "user-interface", "zng"]
 
+[features]
+# Support for loading localization resources from TAR and Tarball.
+tar = ["dep:flate2", "dep:tar"]
+default = ["tar"]
+
 [dependencies]
 zng-ext-l10n-proc-macros = { path = "../zng-ext-l10n-proc-macros", version = "0.2.9" }
 
@@ -36,3 +41,6 @@ once_cell = "1.19"
 dunce = "1.0"
 semver = "1.0"
 serde = { version = "1.0", features = ["derive"] }
+
+flate2 = { version = "1.0", optional = true }
+tar = { version = "0.4", optional = true }
